Alternative octal permissions checker
Assumes that numbers with write or execute bits set but not
read bits are probably bad conversions.

Assumes that user permissions are more generous (i.e. a bigger
number) than group permissions, and that user and group
permissions are more generous than world permissions

There are some false negatives (i.e. numbers that should fail
but succeed such as 4000 which is 07640, but that's not a sensible
permission - all sensible permissions in decimal form (except 440)
contain octal digits 1, 2, or 3 when converted)

dec(440) == oct(670)

which is catered for by the permission generosity order check.