Assure we do not produce duplicated matches (#912)
Due to the way we parse files it is possible to get duplicated
matches.

This change avoids printing duplications and also assures no
regressions are made.

Our examples/example.yml was not used during testing and was a good
example that caused duplicates. Now we test it and assure we
find the exact number of matches that we were expecting.