Refactor root sequence detection in FormattedEmitter (#1964)
* Refactor root sequence detection in FormattedEmitter

In reviewing the code, I found a better place to inject the root_is_sequence detection.
So, move from extending expect_node to expect_document_root.

No change in behavior. Just a bit cleaner.

* Refactor empty_flow_map detection in FormattedEmitter

In reviewing the code, it makes more sense to reset self._in_empty_flow_map
in the same block where it gets set.

I haven't seen anywhere that this actually hits a bug, so it's just a logical inconsistency.