Improve cli argument handling (#2099)
* Include opt-in rules when listing tags and rules

With PR #1450 optional rules with the 'opt-in' tag were introduced
and according to the docs, listing rules and tags should also list
the opt-in rules.

Fixes: #2068

Signed-off-by: Daniel Ziegenberg <daniel@ziegenberg.at>

* gracefully handle invalid format options with listing rules

Signed-off-by: Daniel Ziegenberg <daniel@ziegenberg.at>

* put arguments -L and -T into a mutually exclusive group

We either print a list of rules or a list of tags, but never
both at the same time. So it makes no sense to allow giving
both arguments at the same time.

Signed-off-by: Daniel Ziegenberg <daniel@ziegenberg.at>

* always provide a long form of cli arguments

Having short cli arguments for often used options is nice to have
when using the command in an interactive shell. For writing scripts
it would be better to have long forms of arguments as it increases
the readebility.

Signed-off-by: Daniel Ziegenberg <daniel@ziegenberg.at>

* Update tox.yml

Co-authored-by: Sorin Sbarnea <sorin.sbarnea@gmail.com>