Improve RPM Packaging (#2745)
* rpm: Don't manually specify generated dependencies

%pyproject_buildrequires (opt in) generates build dependencies after
%prep and the pythondist dependency generator (runs by default)
generates runtime dependencies at the end of the rpmbuild process. These
generators remove the need to keep track of dependencies in two places.
%pyproject_buildrequires ensures that the dependencies that the runtime
dependency generator will generate at the end of the build are actually
available.

* Relax dependencies to support Fedora 37 and above

This project aggressively pins dependencies to very recent versions that
aren't available in Fedora. This change relaxes them enough to support
Fedora 37 and Rawhide.

- ansible-compat: 2.0.4 on f36, 2.2.4 on f37, 2.2.5 in f37
  updates-testing, 2.2.5 on rawhide
- filelock: 3.3.1 on f36, 3.7.1 on f37, 3.7.1 on rawhide
- jsonschema: 3.2.0 on f36, 4.10.0 on f37, 4.17.0 on rawhide
- wcmatch: 8.1.2 on f36, 8.1.2 on f37, 8.4.1 on rawhide

* rpm: Use git-core instead of git

git pulls in perl, git-core-docs, and other unneeded dependencies.
git-core is the portion that's actually needed here.

* rpm: Use %pyproject_save_files

%pyproject_save_files generates a file list that should be used instead
of manually specifying the top level module(s) and dist-info
directories.

* rpm: Use %pytest and %pyproject_check_import

Use the %pytest macro to run tests and condense its invocation. It
already sets $PYTHONPATH to the destination site-packages directories.

%pyproject_check_import preforms a basic "smoke test" by importing all
public modules. It's not a replacement for unit tests, but it's better
than nothing when the tests can't run in a network-less mock chroot. The
modules that import pytest are excluded.

* rpm: Fix license and doc handling

- Both GPL-3.0-or-later and MIT apply to the code, so the License tag is
  adjusted accordingly.

- The MIT license text is marked with %license.

- The docs/* glob is changed to docs/. The docs directory has a
  README.md file that replaces the package's actual README.md when the
  structure is flattened.

* rpm: Fix incorrectly formatted %changelog.

Packit will generate a %changelog for you. Remove the incorrectly
formatted placeholder one.